refactor for HOT-2031 Tasks to support changing Ferb endpoint for creating Participants from client ID
