+++ src/main/java/gov/ca/cwds/rest/services/screeningparticipant/OtherAdultInPlacemtHomeTransformer.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.rest.api.services.screeningparticipant;
+package gov.ca.cwds.rest.services.screeningparticipant;
 
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
@@ -12,43 +12,40 @@ import org.joda.time.DateTime;
 
 import gov.ca.cwds.data.legacy.cms.entity.PlacementHome;
 import gov.ca.cwds.data.persistence.cms.CmsPersistentObject;
-import gov.ca.cwds.data.persistence.cms.OtherChildInPlacemtHome;
+import gov.ca.cwds.data.persistence.cms.OtherAdultInPlacemtHome;
 import gov.ca.cwds.rest.api.domain.AddressIntakeApi;
 import gov.ca.cwds.rest.api.domain.IntakeCodeCache;
 import gov.ca.cwds.rest.api.domain.LegacyDescriptor;
 import gov.ca.cwds.rest.api.domain.ParticipantIntakeApi;
 import gov.ca.cwds.rest.api.domain.cms.LegacyTable;
-import gov.ca.cwds.rest.services.submit.Gender;
 
 /**
+ * Business layer object to transform a {@link OtherAdultInPlacemtHome} to an
+ * {@link ParticipantIntakeApi}
+ * 
  * @author CWDS API Team
  *
  */
-public class OtherChildInPlacemtHomeTransformer implements ParticipantMapper {
+public class OtherAdultInPlacemtHomeTransformer implements ParticipantMapper {
 
   @Override
   public ParticipantIntakeApi tranform(CmsPersistentObject object) {
-    OtherChildInPlacemtHome otherChildInPlacemtHome = (OtherChildInPlacemtHome) object;
+    OtherAdultInPlacemtHome otherAdultInPlacemtHome = (OtherAdultInPlacemtHome) object;
 
     LegacyDescriptor otherChildLegacyDescriptor =
-        new LegacyDescriptor(otherChildInPlacemtHome.getId(), null,
-            new DateTime(otherChildInPlacemtHome.getLastUpdatedTime()),
-            LegacyTable.CHILD_IN_PLACEMENT_HOME.getName(),
-            LegacyTable.CHILD_IN_PLACEMENT_HOME.getDescription());
+        new LegacyDescriptor(otherAdultInPlacemtHome.getId(), null,
+            new DateTime(otherAdultInPlacemtHome.getLastUpdatedTime()),
+            LegacyTable.ADULT_IN_PLACEMENT_HOME.getName(),
+            LegacyTable.ADULT_IN_PLACEMENT_HOME.getDescription());
 
-    String firstName = StringUtils.isNotBlank(otherChildInPlacemtHome.getName())
-        ? setFirstName(otherChildInPlacemtHome)
+    String firstName = StringUtils.isNotBlank(otherAdultInPlacemtHome.getName())
+        ? setFirstName(otherAdultInPlacemtHome)
         : null;
-    String lastName = StringUtils.isNotBlank(otherChildInPlacemtHome.getName())
-        ? setLastName(otherChildInPlacemtHome)
-        : otherChildInPlacemtHome.getName();
-    String gender =
-        StringUtils.isNotBlank(otherChildInPlacemtHome.getGenderCode())
-            ? (Gender.findByCmsDescription(otherChildInPlacemtHome.getGenderCode().toUpperCase()))
-                .getNsDescription()
-            : null;
+    String lastName = StringUtils.isNotBlank(otherAdultInPlacemtHome.getName())
+        ? setLastName(otherAdultInPlacemtHome)
+        : otherAdultInPlacemtHome.getName();
 
-    PlacementHome placementHome = otherChildInPlacemtHome.getPlacementHome();
+    PlacementHome placementHome = otherAdultInPlacemtHome.getPlacementHome();
     String streetAddress = placementHome.getStreetNo() + " " + placementHome.getStreetNm();
     String state =
         IntakeCodeCache.global().getIntakeCodeForLegacySystemCode(placementHome.getStateCode());
@@ -65,20 +62,21 @@ public class OtherChildInPlacemtHomeTransformer implements ParticipantMapper {
     addresses = Collections.unmodifiableSet(addresses);
 
     return new ParticipantIntakeApi(null, null, null, otherChildLegacyDescriptor, firstName, null,
-        lastName, null, gender, null, null, null, otherChildInPlacemtHome.getBirthDate(),
-        new LinkedList<>(), null, null, null, new HashSet<>(), addresses, null, false, false);
+        lastName, null, otherAdultInPlacemtHome.getGenderCode(), null, null, null,
+        otherAdultInPlacemtHome.getBirthDate(), new LinkedList<>(), null, null, null,
+        new HashSet<>(), addresses, null, false, false);
   }
 
-  private String setFirstName(OtherChildInPlacemtHome otherChildInPlacemtHome) {
-    String fullName = otherChildInPlacemtHome.getName();
+  private String setFirstName(OtherAdultInPlacemtHome otherAdultInPlacemtHome) {
+    String fullName = otherAdultInPlacemtHome.getName();
     if (!fullName.contains(" ")) {
       return null;
     }
     return StringUtils.split(fullName, " ", 2)[0];
   }
 
-  private String setLastName(OtherChildInPlacemtHome otherChildInPlacemtHome) {
-    String fullName = otherChildInPlacemtHome.getName();
+  private String setLastName(OtherAdultInPlacemtHome otherAdultInPlacemtHome) {
+    String fullName = otherAdultInPlacemtHome.getName();
     if (!fullName.contains(" ")) {
       return fullName;
     }
