+++ src/main/java/gov/ca/cwds/rest/services/screeningparticipant/ParticipantDaoFactory.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.rest.api.services.screeningparticipant;
+package gov.ca.cwds.rest.services.screeningparticipant;
 
 import gov.ca.cwds.data.CrudsDao;
 import gov.ca.cwds.data.persistence.cms.CmsPersistentObject;
