+++ src/main/java/gov/ca/cwds/rest/services/screeningparticipant/ParticipantDaoFactoryImpl.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.rest.api.services.screeningparticipant;
+package gov.ca.cwds.rest.services.screeningparticipant;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
