+++ src/main/java/gov/ca/cwds/rest/services/screeningparticipant/ParticipantMapper.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.rest.api.services.screeningparticipant;
+package gov.ca.cwds.rest.services.screeningparticipant;
 
 import gov.ca.cwds.data.persistence.cms.CmsPersistentObject;
 import gov.ca.cwds.rest.api.domain.ParticipantIntakeApi;
