+++ src/main/java/gov/ca/cwds/rest/services/screeningparticipant/ParticipantMapperFactory.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.rest.api.services.screeningparticipant;
+package gov.ca.cwds.rest.services.screeningparticipant;
 
 /**
  * @author CWDS API Team
